Development

+++ chatroom/auto_follow_members.go
@@ -6,14 +6,14 @@ import (
 	"github.com/nateshr/likeminds-authentication/utils"
 )
 
-//AutoFollowMembersRequest
+// AutoFollowMembersRequest
 type AutoFollowMembersRequest struct {
-	ChatroomId          int32 `json:"chatroom_id" binding:"required"`
+	ChatroomId          int64 `json:"chatroom_id" binding:"required"`
 	AutoFollowDone      *bool `json:"auto_follow_done" binding:"required"`
 	IncludeMembersLater *bool `json:"include_members_later" binding:"required"`
 }
 
-//AutoFollowMembers is used to enable auto follow members for a chatroom
+// AutoFollowMembers is used to enable auto follow members for a chatroom
 func AutoFollowMembers(c *gin.Context) {
 
 	//Authorize User
